use an adjusted wald CIs on success rates and improvements
The adjusted wald interval has much better coverage than the ordinary
wald interval we had been using. it has the unintuitive quality that
the estimated success rate is not numSuccesses / numTrials, but a)
this effect disappears for large samples anb b) I think the adjusted
estimates are more reasonable for small samples. we use the lower
"trial" confidence level to obtain success rate estimates and the
higher confidence level only to compute the interval width on the
improvement. this assures we get reasonable intervals on the success
rates but still end up with an improvement CI of the desired
confidence level.